Convert WindowManager to Class
Typescript likes classes way more that functions that return
objects with methods.
